build: fix mxml linking aginst version 4
There were to issues with the previous commit addressing mxml4 support:

* -lmxml is always appended to the link flags, but mxml4 needs -lmxml4.
* $(MXML_LFLAGS) is appended to the link flags, but is empty (actually,
  $(MXML_LIBS) was meant).

This causes build errors with only mxml4 on the system and includes
mxml4 headers, but links against older mxml versions when both are
present. The only consistent and successful build is with only old mxml
versions installed on the system.

Closes: https://github.com/a-tze/fuse-ts/issues/20